fix for teensy

+++ arduino/panel_pixel_test/panel_pixel_test.ino
@@ -42,14 +42,14 @@ extern unsigned int __bss_end;
 extern unsigned int __heap_start;
 extern void *__brkval;
 
-uint16_t getFreeSram() {
+int getFreeSram() {
   uint8_t newVariable;
   // heap is empty, use bss as start memory address
-  if ((uint16_t)__brkval == 0)
-    return (((uint16_t)&newVariable) - ((uint16_t)&__bss_end));
+  if ((int)__brkval == 0)
+    return (((int)&newVariable) - ((int)&__bss_end));
   // use heap end as the start of the memory address
   else
-    return (((uint16_t)&newVariable) - ((uint16_t)__brkval));
+    return (((int)&newVariable) - ((int)__brkval));
 };
 
 #define SNPRINTLN(...) \
